Now use pushState to update the URL
When the AJAX request fires to store the snippet, the server responds
with the ID for the new page. This commit fetches that ID and uses
pushState to amend the URL to the one for that ID without reloading the
page. If the user subsequently refreshes the page they'll see the same
content.